% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASSOCShiny.R
\name{ASSOCShiny}
\alias{ASSOCShiny}
\title{Launch 'ASSOCShiny' Interface}
\usage{
ASSOCShiny()
}
\value{
Nothing
}
\description{
ASSOCShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start ASSOCShiny

The interactive user interface is to provide an easy way for association rule mining analysis and downloading relevant plots.
}
\examples{
if(interactive()){
library(rmarkdown)
ASSOCShiny()
}
}
\keyword{ASSOCShiny}
